+++ pkg/registry/registry.go
@@ -25,6 +25,7 @@ import (
 	"github.com/NVIDIA/topograph/pkg/engines/slurm"
 	"github.com/NVIDIA/topograph/pkg/providers"
 	"github.com/NVIDIA/topograph/pkg/providers/aws"
+	"github.com/NVIDIA/topograph/pkg/providers/crusoe"
 	"github.com/NVIDIA/topograph/pkg/providers/dra"
 	"github.com/NVIDIA/topograph/pkg/providers/dsx"
 	"github.com/NVIDIA/topograph/pkg/providers/gcp"
@@ -40,6 +41,8 @@ import (
 var Providers = providers.NewRegistry(
 	aws.NamedLoader,
 	aws.NamedLoaderSim,
+	crusoe.NamedLoader,
+	crusoe.NamedLoaderSim,
 	infiniband.NamedLoaderBM,
 	infiniband.NamedLoaderK8S,
 	dra.NamedLoader,